feat: implement RunwayMonthGridBlock component and integrate it into page rendering
