+++ website/src/components/content-types/page.tsx
@@ -27,6 +27,7 @@ import type { Page } from '@/generated/storyblok/types/109655/storyblok-componen
 import type { WebsiteLanguage, WebsiteRegion } from '@/lib/i18n/utils';
 import type { ParsedUrlQueryInput } from 'querystring';
 import { Fragment } from 'react';
+import { RunwayMonthGridBlock } from '../content-blocks/runway-month-grid-block';
 
 type PageBlock = Page['content'][number];
 type RichtextButtonHeaderAction = 'createProgram';
@@ -100,7 +101,7 @@ const renderPageBlock = (
 		case 'videoText':
 			return <VideoTextBlock blok={block} />;
 		case 'runwayMonthGrid':
-			return <>blabla</>;
+			return <RunwayMonthGridBlock blok={block} lang={lang} />;
 		default:
 			block satisfies never;
